+++ jest.config.js
@@ -37,6 +37,8 @@ module.exports = {
     "src/app/features/hybrid-keymanager/unwrap.service.ts",
     "src/app/features/hybrid-keymanager/sign.service.ts",
     "src/app/features/hybrid-keymanager/sign.api.ts",
+    "src/app/core/services/activity.service.ts",
+    "src/app/features/activity/activity.page.ts",
   ],
   coveragePathIgnorePatterns: [
     '<rootDir>/node_modules/',